software/server_final/requirements.txt
Flask==3.0.2
paho-mqtt==1.6.1
python-dotenv==1.0.1